Created the getDriver class
This class is a static class.  When driver is called this class checks
whether it is already open.  If not a new driver is kicked off from this
static class.  There are currently errors as for it to work we need to
add a config file and a constant class - used to type name of browser
which in turn goes through switch statement.