Added Master Hook class
Unfortunately when running I unable to run browser exception null.  This
has been narrowed to something in the readconfig class.  This current
version has commented out the call on the DriverFactory class to the
readconfig class.  Instead it just uses String variable browserName to
use the chrome value for the switch statement.